PARTE_2/requirements.txt
autopep8==1.6.0
click==8.0.4
dnspython==2.2.1
Flask==2.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pycodestyle==2.8.0
PyJWT==2.3.0
pymongo==4.0.2
python-dotenv==0.19.2
toml==0.10.2
Werkzeug==2.0.3
